requirements.txt
Django
python-decouple
dj_database_url
requests
whitenoise
django-allauth[socialaccount]
dropbox